requirements.txt
scipy==1.2.1
pandas==0.24.2
plotly==3.8.1
h5py==2.9.0
matplotlib==3.1.0
numpy==1.16.3
mdtraj==1.9.2
biopython==0.1.0
pymol==0.1.0
